Switch cache count to long

+++ tests/Headless.DistributedLocks.Tests.Unit/Fakes/FakeDistributedLockStorage.cs
@@ -64,7 +64,7 @@ internal sealed class FakeDistributedLockStorage : IDistributedLockStorage
         return ValueTask.FromResult<IReadOnlyDictionary<string, string>>(result);
     }
 
-    public ValueTask<int> GetCountAsync(string prefix = "")
+    public ValueTask<long> GetCountAsync(string prefix = "")
     {
         var count = string.IsNullOrEmpty(prefix)
             ? _locks.Count
